==> 00_Initialization.R <==

##~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
##                                                                            --
## THIS ENVIRONMENT INITIALIZATION SCRIPT WILL INSTALLS ALL REQUIRED PACKAGE----
##                                                                            --
##~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~

install.packages("tidyverse")
install.packages("janitor")
install.packages("here")
install.packages("lubridate")
install.packages("ggplot2")
install.packages("patchwork")
install.packages("paletteer")
install.packages("DiagrammeR")
